Brainstorm for different titles
Hi 弘哲, I found your tutorial on YouTube, which is the best one on the internet in my opinion. I bookmarked this repo for my future reference. While I was reading it, I came up with ideas of making a couple of titles more clear. Please let me know if the proposal makes sense for you. 

Nice job creating this repo and the videos.

Regards,
Teddy